cmd/storj-sim: Use --defaults=dev by default enabling release tag testing.
Before this change storj-sim would fail to work on a release tag because
of the behavior of --defaults in the presence of an actual release
version file private/version/release.go.

It is desirable for storj-sim (and therefor go test) to work on an
unmodified checkout of the release tag. This makes that possible.

SM-159

Change-Id: I744974973c2b6aca96d0e2965ce140948d5bfaa2